Mitm by Bing (rev. 0.2)
A mitm script in python, first version generated by chatgpt (3.5) with the correct promt engeneering  it became a party functional script, in that proces i noticed funny behavior on the side of chatgpt, namely the arp poisining functional was done in such a way that it seemed to perform the function, but in reality chatgpt made it  look so but it was not doing anyform of arp poisining, so basicly chatgpt lied. I put that in a prompt back to chatgpt, then he made it send only 2 arp packeges,  further prompting led in an functional script. Than i took that script and copied it in bing chat (gpt4 version) and ask him to correct it and make it better, so hé made it more complex, plus i ask to come up with functions to make it perform better, hè did this in the forum of ai intrussion detection systeem. This is the draft i got from Bing.  I going to test this, and then look how much further Bing chat can bring this.